Move new scatter code into place
