+++ glue/viewers/scatter/layer_style_editor.ui
