+++ glue/viewers/scatter/qt/__init__.py
