+++ glue/viewers/scatter/qt/layer_style_editor.py
