+++ glue/viewers/scatter/qt/options_widget.ui
