+++ glue/viewers/scatter/qt/tests/test_viewer_widget.py
